DTNHost: Revert sat.-specific changes, allow jumping to first location
This removes the dependency upon SatelliteMovement. Furthermore, a
static setting is introduced which enables the host to "jump" directly
to the first location on each new path retrieved from the movement
class. Thus, if a new path is provided and the first node on it is not
equal to the current position (which should not be the case in _most_
scenarios), the node jumps directly to this location. This enables maps
which are transparently "endless", i.e. allow nodes to appear on the
left side when they disappear on the right side. Such a behavior is e.g.
needed in a satellite scenario if projecting the world map on the ONE
map.