MovementModel: Add setting for "jumping" to initial locations
If a new path is provided, by default nodes are moved smoothly to the
start location. This may be not wanted if the first point on the new
path is at a completely different location, e.g. for the 2D satellite
movement.

This adds a setting "MovementModel.jumpToInitialPathLocations" to allow
enabling a "jump" to the first point on the new path each time a new
path is fetched.